+++ discovery-provider/src/api/v1/api.py
@@ -9,7 +9,6 @@ from src.api.v1.metrics import ns as metrics_ns
 from src.api.v1.search import full_ns as full_search_ns
 from src.api.v1.models.users import ns as models_ns
 from src.api.v1.resolve import ns as resolve_ns
-from src.api.v1.challenge import ns as challenge_ns
 
 
 class ApiWithHTTPS(Api):
@@ -32,7 +31,6 @@ api_v1.add_namespace(tracks_ns)
 api_v1.add_namespace(challenges_ns)
 api_v1.add_namespace(metrics_ns)
 api_v1.add_namespace(resolve_ns)
-api_v1.add_namespace(challenge_ns)
 
 bp_full = Blueprint("api_v1_full", __name__, url_prefix="/v1/full")
 api_v1_full = ApiWithHTTPS(bp_full, version="1.0")
